Corrected a bug in a memory write function. Thanks to drk||Razi for spotting this one.
git-svn-id: http://pcsx2-playground.googlecode.com/svn/trunk@156 a6443dda-0b58-4228-96e9-037be469359c

+++ pcsx2/Hw.c
@@ -1066,7 +1066,10 @@ void hwWrite32(u32 mem, u32 value) {
 			}
 */
 			// omg, just xor it
-			psHu32(0xf010) ^= value;
+			//psHu32(0xf010) ^= value;
+			//drk says its wrong (updates all 32bits instead of 16 only 
+			//so lets use this and see :) (rama)
+			psHu32(0xf010) ^= (u16)value;
 
 			if ((cpuRegs.CP0.n.Status.val & 0x10407) == 0x10401)
                 cpuTestINTCInts();
